+++ mplapack/reference/Clansb.cpp
@@ -76,7 +76,7 @@ REAL Clansb(const char *norm, const char *uplo, INTEGER const n, INTEGER const k
         value = zero;
         if (Mlsame(uplo, "U")) {
             for (j = 1; j <= n; j = j + 1) {
-                for (i = max(k + 2 - j, 1); i <= k + 1; i = i + 1) {
+                for (i = max(k + 2 - j, (INTEGER)1); i <= k + 1; i = i + 1) {
                     sum = abs(ab[(i - 1) + (j - 1) * ldab]);
                     if (value < sum || Risnan(sum)) {
                         value = sum;
@@ -149,7 +149,7 @@ REAL Clansb(const char *norm, const char *uplo, INTEGER const n, INTEGER const k
                 for (j = 2; j <= n; j = j + 1) {
                     colssq[1 - 1] = zero;
                     colssq[2 - 1] = one;
-                    Classq(min(j - 1, k), &ab[(max(k + 2 - j, 1) - 1) + (j - 1) * ldab], 1, colssq[1 - 1], colssq[2 - 1]);
+                    Classq(min(j - 1, k), &ab[(max(k + 2 - j, (INTEGER)1) - 1) + (j - 1) * ldab], 1, colssq[1 - 1], colssq[2 - 1]);
                     Rcombssq(ssq, colssq);
                 }
                 l = k + 1;
